Write tests for RandomSet accessors
Adds tests, where we try to both add an already existing element, and
delete a non-existing one. Both times the internal contents should not
change.

Also adds a test where we try to access a random element, which should
be in the same range, that the set was filled from.